Made KernelTransaction implement TxTermination
Commit removes TxTermination implementation and makes KernelTransaction
implement this interface directly. This results in less indirections during
`#shouldBeTerminated()` call and fits nicely in KernelTransaction interface
which already had all termination related methods.
