+++ community/kernel/src/main/java/org/neo4j/kernel/impl/api/KernelTransactionImplementation.java
@@ -53,7 +53,6 @@ import org.neo4j.kernel.impl.api.state.TxState;
 import org.neo4j.kernel.impl.api.store.PersistenceCache;
 import org.neo4j.kernel.impl.api.store.StoreReadLayer;
 import org.neo4j.kernel.impl.api.tx.TxTermination;
-import org.neo4j.kernel.impl.api.tx.TxTerminationImpl;
 import org.neo4j.kernel.impl.index.IndexEntityType;
 import org.neo4j.kernel.impl.locking.LockGroup;
 import org.neo4j.kernel.impl.locking.Locks;
@@ -157,7 +156,7 @@ public class KernelTransactionImplementation implements KernelTransaction, TxSta
     private Locks.Client locks;
     private boolean closing, closed;
     private boolean failure, success;
-    private final TxTerminationImpl termination = new TxTerminationImpl();
+    private volatile boolean terminated;
     // Some header information
     private long startTimeMillis;
     private long lastTransactionIdWhenStarted;
@@ -219,7 +218,7 @@ public class KernelTransactionImplementation implements KernelTransaction, TxSta
         this.locks = newLocksClient();
         this.context.bind( locks );
         this.closing = closed = failure = success = false;
-        this.termination.reset();
+        this.terminated = false;
         this.transactionType = TransactionType.ANY;
         this.beforeHookInvoked = false;
         this.recordState.initialize( lastCommittedTx );
@@ -245,16 +244,16 @@ public class KernelTransactionImplementation implements KernelTransaction, TxSta
     @Override
     public boolean shouldBeTerminated()
     {
-        return termination.shouldBeTerminated();
+        return terminated;
     }
 
     @Override
     public void markForTermination()
     {
-        if ( !termination.shouldBeTerminated() )
+        if ( !terminated )
         {
             failure = true;
-            termination.markForTermination();
+            terminated = true;
             transactionMonitor.transactionTerminated();
         }
     }
@@ -978,7 +977,7 @@ public class KernelTransactionImplementation implements KernelTransaction, TxSta
     {
         if ( TX_TERMINATION_AWARE_LOCKS )
         {
-            return locksManager.newClient( termination );
+            return locksManager.newClient( this );
         }
         return locksManager.newClient( TxTermination.NONE );
     }
